Feature/discretization

+++ src/types.rs
@@ -1,3 +1,3 @@
 //! A common place to define types used in the pipeline
 
-pub type Numeric = f32;
+pub type Numeric = f64;
